Refactor: controller decoupling
