+++ server/controllers/decorators/index.ts
@@ -1,0 +1,4 @@
+import { controller } from './controller';
+import { use } from './use';
+
+export { controller, use };
