+++ server/interfaces/customRequest.ts
@@ -1,0 +1,5 @@
+import { Request } from 'express';
+
+export interface customRequest extends Request {
+  user?: { id: number };
+}
