+++ server/interfaces/index.ts
@@ -1,0 +1,6 @@
+import { userInterface } from './userInterface';
+import { metadataKeys } from './metadataKeys';
+import { methods } from './methods';
+import { customRequest } from './customRequest';
+
+export { userInterface, metadataKeys, methods, customRequest };
