+++ server/interfaces/methods.ts
